thermal: core: Speed up resume by queueing thermal zone updates
Thermal zone updates for all the cores happen in the post suspend notifier
context. This process can take a variable amount of time depending upon the
sensors that are read as a part of update.

To avoid delaying the system resume, queue the thermal zone device update
from the resume context.

Signed-off-by: Ram Chandrasekar <rkumbako@codeaurora.org>
Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>