thermal: core: Create new high priority workqueue
Thermal framework uses the system freezable work queue for processing
governor actions and mitigation action is performed in the same context.
System workqueue allows processing of a maximum of one active event at a
time. If the mitigation action is "slow", it creates a bottleneck for
the rest of the workqueue event processing and results in delayed action
and possible temperature overshoot.

In order to avoid this, create a new high priority thermal workqueue and
move all passive monitoring to this context. This workqueue allows
multiple work events to be processed simultaneously since we allocate
the workqueue with max_active set to 0. As a result, we end up able to
serve up to WQ_DFL_ACTIVE (256) simultaneous events.

Heavily based on downstream patch from Ram Chandrasekar
<rkumbako@codeaurora.org>

Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>