drivers: thermal: do not clobber cooling dev state from userspace
Let userspace be another voter for cooling device state instead of the
overriding authority. It is possible that the thermal governor may find
a higher cooling state desirable than what is recommended by the
userspace. Separate out the current cooling device state from the
userspace request and aggregate the userspace request with the
governors' recommendation.

Signed-off-by: Lina Iyer <ilina@codeaurora.org>
Original-patch-by: Ram Chandrasekar <rkumbako@codeaurora.org>