arm64: dts: qcom: sdm845: Add LMH node
Split the node definition into the generic and the board-specific part
to allow for other boards that might wire up their PMIC differently.

The limits hardware module on the sdm845-mtp is powered by the always-on
vreg_l1a_0p875 regulator.

Add the LMH thermal zone and wireup each cpu to a LMH instance.

Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>
Message-Id: <9f193ff3716a73134994cab33e3c7564882f5115.1589306399.git.amit.kucheria@linaro.org>