of-thermal: Disable polling when interrupt property is found in DT
Currently, in order to enable interrupt-only mode, one must set
polling-delay-passive and polling-delay properties in the DT to 0,
otherwise the thermal framework will continue to setup a periodic timers
to monitor the thermal zones.

Change the behaviour, so that on DT-based systems, we no longer have to
set the properties to zero if we find an 'interrupt' property in the
sensor.

Following data shows the number of times
thermal_zone_device_set_polling() is invoked with and without this
patch. So the patch achieves the same behaviour as setting the delay
properties to 0.

Current behaviour (without setting delay properties to 0):
  FUNC                              COUNT
  thermal_zone_device_update          302
  thermal_zone_device_set_pollin     7911

Current behaviour (with delay properties set to 0):
  FUNC                              COUNT
  thermal_zone_device_update            3
  thermal_zone_device_set_pollin        6

With this patch (without setting delay properties to 0):
  FUNC                              COUNT
  thermal_zone_device_update            3
  thermal_zone_device_set_pollin        6

Suggested-by: Stephen Boyd <swboyd@chromium.org>
Signed-off-by: Amit Kucheria <amit.kucheria@linaro.org>