powercap/drivers/energy_model: protocode: Add powercap energy model based
On the embedded world, the complexity of the SoC leads to an
increasing number of hotspots which need to be monitored and mitigated
as a whole in order to prevent the going above the normative and
legally stated 'skin temperature'.

The recent introduction of the energy model allows to get power
information related to a gpu or a cpu device. The energy model is now
being improved to be generic and extended to all devices.

On the other side, the powercap infrastructure is a perfect fit to define
power constraints in a hierarchical way.

This patch introduces the powercap energy model based.

The first step is to define power constraints at different levels.

Example of hierarchy:

Soc
  |
  |-- gpu
  |
  `-- package
        |
	|-- perfdomain0
	|         |
	|         |-- cpu0
	|         |
	|         |-- cpu1
	|         |
	|         |-- cpu2
	|         |
	|         `-- cpu3
	|
	`-- perfdomain1
	          |
		  |-- cpu4
		  |
		  `-- cpu5

The leaves of the hierarchy are the ones returning the max power range
and the current consumed power.

The intermediate nodes of the tree call recursively their children ops
and sum their power.

For instance: cpu[0-4] have 179mW max, perfdomain0 will have 716mW max.

Each node have a weight on a 2^10 basis, in order to reflect the
percentage of power distribution of the children's node. This
percentage is used to dispatch the power limit to the children.

This simple approach allows to do a fair distribution of the power
limit but it will be replaced by a more complex mechanism where the
power limit will be dynamically adjusted depending on the power
consumption of the siblings.

Signed-off-by: Daniel Lezcano <daniel.lezcano@linaro.org>
Message-Id: <20200529103738.32542-1-daniel.lezcano@linaro.org>