powercap_em: automatic registration of powercap em zones based on genpd
genpd already captures the physical and virtual hierarchy of how blocks
in an SoC are powered and their various dependencies. Use that to
create an equivalent powercap hierarchy.

FIXME: This is proof-of-concept that doesn't take in to account child
and sibling relationships in genpd. Currently, all powercap zones are
created at the root level. This needs to be fixed.

FIXME #2: The energy costing for devices could be looked up from the
energy model during init to fill up the constraint table.

FIXME #3: Revisit if there is a need for a separate property inside
genpd DT bindings to powercap.

Signed-off-by: Amit Kucheria <amitk@kernel.org>